split vpr and genfasm
